[v9_9] misc x64 VS 2015 CTP fixes [#39308]
(cherry picked from commit ab973ec40cc7a79bbf930085b1293824caf73e7f)
(cherry picked from commit 0fdcf394cb0b0fa60a7da9f852478ec20486550b)

+++ lib/dns/rdata.c
@@ -1398,7 +1398,7 @@ multitxt_fromtext(isc_textregion_t *source, isc_buffer_t *target) {
 		if (escape)
 			return (DNS_R_SYNTAX);
 
-		isc_buffer_add(target, t - t0);
+		isc_buffer_add(target, (unsigned int)(t - t0));
 	} while (n != 0);
 	return (ISC_R_SUCCESS);
 }
